+++ framework/validators/Validator.php
@@ -41,6 +41,7 @@ use yii\base\NotSupportedException;
  * - `string`: [[StringValidator]]
  * - `trim`: [[FilterValidator]]
  * - `unique`: [[UniqueValidator]]
+ * - `unsafe`: [[UnsafeValidator]]
  * - `url`: [[UrlValidator]]
  *
  * @author Qiang Xue <qiang.xue@gmail.com>
@@ -80,6 +81,10 @@ class Validator extends Component
             'skipOnArray' => true,
         ],
         'unique' => 'yii\validators\UniqueValidator',
+        'unsafe' => [
+            'class' => 'yii\validators\SafeValidator',
+            'safe' => false,
+        ],
         'url' => 'yii\validators\UrlValidator',
     ];
     /**
@@ -172,6 +177,12 @@ class Validator extends Component
      * @see when
      */
     public $whenClient;
+	/**
+	 * @var boolean whether attributes listed with this validator should be considered safe for massive assignment.
+	 * Defaults to true.
+	 * @since 2.0.4
+	 */
+	public $safe = true;
 
 
     /**